move challenge into src directory
