+++ pyproject.toml
@@ -8,9 +8,7 @@ readme = "README.md"
 repository = "https://github.com/bionikspoon/interview-challenge"
 keywords = ["scavenger hunt", "flask", "brain teaser", "interview challenge"]
 packages = [
-    { include = "challenge_server.py" },
-    { include = "solution/**/*.py" },
-    { include = "migrate.py" },
+    { include = "src/**/*.py" },
 ]
 
 [tool.poetry.dependencies]
@@ -30,6 +28,6 @@ build-backend = "poetry.core.masonry.api"
 
 
 [tool.poetry.scripts]
-migrate = "migrate:run"
-serve-dev = "challenge_server:run"
-solve = "solution.solution:run"
+migrate = "src.migrate:run"
+challenge_server = "src.challenge_server:run"
+solve = "src.solution:run"
