+++ src/wsgi.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # coding=utf-8
 
-from challenge_server import app
+from src.challenge_server import app
 
 if __name__ == "__main__":
     app.run()
